requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   starlette
    #   watchfiles
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via -r requirements.in
google-api-core==2.25.1
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.181.0
    # via -r requirements.in
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.2
    # via -r requirements.in
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==2.19.0
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via google-api-core
h11==0.16.0
    # via uvicorn
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
lxml==6.0.1
    # via python-docx
oauthlib==3.3.1
    # via requests-oauthlib
proto-plus==1.26.1
    # via google-api-core
protobuf==5.29.5
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.9
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pymupdf==1.26.4
    # via -r requirements.in
pyparsing==3.2.4
    # via httplib2
python-docx==1.2.0
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.47.3
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
